Fixed up tests

+++ plugin-server/tests/main/ingestion-queues/session-recording/services/session-manager-v3.test.ts
@@ -2,48 +2,13 @@ import { Upload } from '@aws-sdk/lib-storage'
 import fs from 'fs/promises'
 import { DateTime, Settings } from 'luxon'
 import path from 'path'
+import { PassThrough } from 'stream'
 import * as zlib from 'zlib'
 
 import { defaultConfig } from '../../../../../src/config/config'
 import { SessionManagerV3 } from '../../../../../src/main/ingestion-queues/session-recording/services/session-manager-v3'
 import { now } from '../../../../../src/main/ingestion-queues/session-recording/utils'
 import { createIncomingRecordingMessage } from '../fixtures'
-
-// class MockStream {
-//     write = jest.fn(() => true)
-//     close = jest.fn((cb) => cb?.())
-//     end = jest.fn((cb) => cb?.())
-//     on = jest.fn()
-//     once = jest.fn((val, cb) => cb?.())
-// }
-
-// jest.mock('fs', () => {
-//     return {
-//         ...jest.requireActual('fs'),
-//         writeFileSync: jest.fn(),
-//         createReadStream: jest.fn(() => new MockStream()),
-//         createWriteStream: jest.fn(() => new MockStream()),
-//     }
-// })
-
-// jest.mock('stream/promises', () => {
-//     return {
-//         ...jest.requireActual('stream/promises'),
-//         pipeline: jest.fn(
-//             () =>
-//                 new Promise<void>((_res) => {
-//                     // do nothing
-//                 })
-//         ),
-//     }
-// })
-
-// jest.mock('stream', () => {
-//     return {
-//         ...jest.requireActual('stream'),
-//         PassThrough: jest.fn(() => new MockStream()),
-//     }
-// })
 
 jest.mock('@aws-sdk/lib-storage', () => {
     const mockUpload = jest.fn().mockImplementation(() => {
@@ -58,13 +23,6 @@ jest.mock('@aws-sdk/lib-storage', () => {
         Upload: mockUpload,
     }
 })
-
-// jest.mock('fs/promises', () => {
-//     return {
-//         ...jest.requireActual('fs/promises'),
-//         unlink: jest.fn().mockResolvedValue(undefined),
-//     }
-// })
 
 const tmpDir = path.join(__dirname, '../../../../../.tmp/test_session_recordings')
 
@@ -94,7 +52,9 @@ describe('session-manager', () => {
         // it's always May 25
         Settings.now = () => new Date(2018, 4, 25).valueOf()
 
-        await fs.rmdir(tmpDir, { recursive: true })
+        if (await fs.stat(tmpDir).catch(() => null)) {
+            await fs.rmdir(tmpDir, { recursive: true })
+        }
 
         sessionManager = await createSessionManager()
     })
@@ -123,7 +83,7 @@ describe('session-manager', () => {
             createdAt: 1527202800000,
         })
 
-        const stats = await fs.stat(`${sessionManager.context.dir}/buffer.jsonl.gz`)
+        const stats = await fs.stat(`${sessionManager.context.dir}/buffer.jsonl`)
         expect(stats.size).toBeGreaterThan(0)
     })
 
@@ -139,7 +99,7 @@ describe('session-manager', () => {
         await sessionManager.add(event)
         await sessionManager.flush()
 
-        expect(await fs.readdir(sessionManager.context.dir)).toEqual(['buffer.jsonl.gz', 'metadata.json'])
+        expect(await fs.readdir(sessionManager.context.dir)).toEqual(['buffer.jsonl', 'metadata.json'])
     })
 
     it('does flush if the stored file is older than the threshold', async () => {
@@ -169,7 +129,7 @@ describe('session-manager', () => {
         expect(mockUploadCalls[0][0]).toEqual(
             expect.objectContaining({
                 params: expect.objectContaining({
-                    Key: `session_recordings/team_id/1/session_id/session_id_1/data/${firstTimestamp}-${lastTimestamp}.jsonl.gz`,
+                    Key: `session_recordings/team_id/1/session_id/session_id_1/data/${firstTimestamp}-${lastTimestamp}.jsonl`,
                 }),
             })
         )
@@ -187,16 +147,16 @@ describe('session-manager', () => {
     it('not remove files when stopped', async () => {
         expect(await fs.readdir(sessionManager.context.dir)).toEqual([])
         await sessionManager.add(createIncomingRecordingMessage())
-        expect(await fs.readdir(sessionManager.context.dir)).toEqual(['buffer.jsonl.gz', 'metadata.json'])
+        expect(await fs.readdir(sessionManager.context.dir)).toEqual(['buffer.jsonl', 'metadata.json'])
         await sessionManager.stop()
-        expect(await fs.readdir(sessionManager.context.dir)).toEqual(['buffer.jsonl.gz', 'metadata.json'])
+        expect(await fs.readdir(sessionManager.context.dir)).toEqual(['buffer.jsonl', 'metadata.json'])
     })
 
     it('removes the directly when stopped after fully flushed', async () => {
         const sm = await createSessionManager('session_id_2', 2, 2)
         expect(await fs.readdir(sm.context.dir)).toEqual([])
         await sm.add(createIncomingRecordingMessage())
-        expect(await fs.readdir(sm.context.dir)).toEqual(['buffer.jsonl.gz', 'metadata.json'])
+        expect(await fs.readdir(sm.context.dir)).toEqual(['buffer.jsonl', 'metadata.json'])
         await sm.flush(true)
         expect(await fs.readdir(sm.context.dir)).toEqual([])
         await sm.stop()
@@ -214,7 +174,7 @@ describe('session-manager', () => {
         })
         await sessionManager.add(event)
 
-        const content = await fs.readFile(`${sessionManager.context.dir}/buffer.jsonl.gz`, 'utf-8')
+        const content = await fs.readFile(`${sessionManager.context.dir}/buffer.jsonl`, 'utf-8')
         expect(content).toEqual(
             '{"window_id":"window_id_1","data":[{"timestamp":170000000,"type":4,"data":{"href":"http://localhost:3001/"}},{"timestamp":170001000,"type":4,"data":{"href":"http://localhost:3001/"}}]}\n'
         )
@@ -242,12 +202,48 @@ describe('session-manager', () => {
             })
         )
 
-        const buffer = await fs.readFile(`${sm1.context.dir}/buffer.jsonl.gz`, 'utf-8')
+        const buffer = await fs.readFile(`${sm1.context.dir}/buffer.jsonl`, 'utf-8')
         expect(buffer).toEqual(
             '{"window_id":"window_id_1","data":[{"timestamp":170000000,"type":4,"data":{"href":"http://localhost:3001/"}},{"timestamp":170001000,"type":4,"data":{"href":"http://localhost:3001/"}}]}\n{"window_id":"window_id_1","data":[{"timestamp":170002000,"type":4,"data":{"href":"http://localhost:3001/"}},{"timestamp":170003000,"type":4,"data":{"href":"http://localhost:3001/"}}]}\n'
         )
 
         await sm1.stop()
         await sm2.stop()
+    })
+
+    it('uploads a gzip compressed file to S3', async () => {
+        await sessionManager.add(
+            createIncomingRecordingMessage({
+                events: [{ timestamp: 170000000, type: 4, data: { href: 'http://localhost:3001/' } }],
+            })
+        )
+        await sessionManager.flush(true)
+
+        const mockUploadCalls = (Upload as unknown as jest.Mock).mock.calls
+        console.log(mockUploadCalls[0][0])
+        expect(mockUploadCalls[0][0]).toMatchObject({
+            params: {
+                Bucket: 'posthog',
+                Key: 'session_recordings/team_id/1/session_id/session_id_1/data/170000000-170000000.jsonl',
+                ContentEncoding: 'gzip',
+                ContentType: 'application/json',
+            },
+        })
+        const uploadBody = mockUploadCalls[0][0].params.Body
+
+        expect(uploadBody).toBeInstanceOf(PassThrough)
+
+        // Extract the content from the stream and gzip decompress it
+        const chunks: Uint8Array[] = []
+        for await (const chunk of uploadBody) {
+            chunks.push(chunk)
+        }
+
+        const buffer = Buffer.concat(chunks)
+        const uncompressed = zlib.gunzipSync(buffer).toString('utf-8')
+
+        expect(uncompressed).toEqual(
+            '{"window_id":"window_id_1","data":[{"timestamp":170000000,"type":4,"data":{"href":"http://localhost:3001/"}}]}\n'
+        )
     })
 })
